+++ templates/environments/test-environment-template.js
@@ -1,5 +1,4 @@
 function makeTestEnvironment(simpleMock) {
-  // CouchDB utility functions
   const isArray = Array.isArray;
   const log = simpleMock.stub();
   const toJSON = JSON.stringify;
@@ -13,6 +12,21 @@ function makeTestEnvironment(simpleMock) {
     log,
     sum,
     toJSON,
-    validationFunction: $VALIDATION_FUNC_PLACEHOLDER$
+    validationFunction: function(newDoc, oldDoc, userContext, securityInfo) {
+      try {
+        ($VALIDATION_FUNC_PLACEHOLDER$)(newDoc, oldDoc, userContext, securityInfo);
+      } catch (error) {
+        /* Ensure that validation function errors are thrown as Error objects in test cases to avoid misleading test
+           failure messages (https://github.com/OldSneerJaw/couchster/issues/15) */
+        if (error.forbidden && !(error instanceof Error)) {
+          const wrapperError = new Error(error.forbidden);
+          wrapperError.forbidden = error.forbidden;
+
+          throw wrapperError;
+        } else {
+          throw error;
+        }
+      }
+    }
   };
 }